+++ eol_scons/tool.py
@@ -10,6 +10,7 @@ functionality which applies the eol_scons extensions as if it were a tool.
 
 import os
 import re
+from pathlib import Path
 import SCons.Tool
 from SCons.Script.SConscript import global_exports
 from SCons.Script import COMMAND_LINE_TARGETS
@@ -312,6 +313,20 @@ def Require(env, tools):
     return applied
 
 
+_scripts_dir = None
+
+
+def ScriptsDir() -> str:
+    """
+    Return absolute path to eol_scons scripts directory.
+    """
+    global _scripts_dir
+    if _scripts_dir is None:
+        pscripts = (Path(__file__).parent / "../scripts").resolve()
+        _scripts_dir = str(pscripts.resolve())
+    return _scripts_dir
+
+
 def generate(env, **_kw):
     """
     Generate the basic eol_scons customizations for the given environment,
@@ -324,6 +339,8 @@ def generate(env, **_kw):
         env.LogDebug("skipping _generate(), already applied")
         return
     env._eol_scons_generated = True
+
+    env['EOL_SCONS_SCRIPTS_DIR'] = ScriptsDir()
 
     # Setup methods
     eol_scons.methods.AddMethods(env)